Formatted the final writeup, created the pdf file
I changed the input parameters and all to simple file names in the
decTree.py. It is too complex and not needed.
Finished with final formatting and created the pdf.
This is ready for submission. Just check the filenames in the
decTree.py.
For the foodInspection do we need foodInspection pruned or just
foodinspection.
Just check that, run the program once and we are done.

Make the final submission when you please.